LSN-177In edit Stakeholder empty select box coming with label SelectBox CLSN-178Save Button not getting enabled in edit stakeholder mode and delete button of payment not working CLSN-189radio button binding issue

+++ UserInterface/Generic/permissions/PermissionScript.js
@@ -78,10 +78,12 @@
     var setUpdatesinPrem = function (hierarchies) {
         dldata.permissionsChanged = false;
         _.forEach(hierarchies, function (item) {
-            var perm = JSON.parse(item.Permissions);
-            deleteOld(perm, permFactory.permission);
-            addNew(perm, permFactory.permission);
-            item.Permissions = JSON.stringify(perm);
+            if (!$csfactory.isEmptyObject(item.Permissions)) {
+                var perm = JSON.parse(item.Permissions);
+                deleteOld(perm, permFactory.permission);
+                addNew(perm, permFactory.permission);
+                item.Permissions = JSON.stringify(perm);
+            }
         });
     };
